feat: integrate Supermemory and fix client-side build error
- Integrated Supermemory AI SDK for persistent agent memory.
- Refactored `lib/utils/index.ts` to separate server-only AI logic into `lib/utils/ai.ts`.
- This fixes a build error where client components (like `chat.tsx`) were indirectly importing `next/headers`.
- Updated all agents and actions to use the new server-only utility for model retrieval.
- Propagated `userId` and `chatId` to all agents for user-scoped memory.
